# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ruian-backend
  description: "\n    ## 锐安后端 API 文档\n\n    本 API 提供以下功能：\n\n    * **认证管理**: 用户登录和身份验证\n\
    \    * **文件上传**: 支持本地存储和阿里云 OSS 两种方式上传 Excel 文件\n    * **生产信息查询**: 查询和管理生产订单信息\n\
    \    * **员工工作量查询**: 查询和管理员工工作量记录\n    * **数据核验**: 验证生产信息和员工工作量数据的一致性\n\n    ##\
    \ 认证说明\n\n    大部分接口需要 Bearer Token 认证。请先调用 `/api/auth/login` 接口获取 token，\n   \
    \ 然后在请求头中添加：`Authorization: Bearer <your_token>`\n\n    ## 错误响应\n\n    所有错误响应都遵循统一格式：\n\
    \    ```json\n    {\n        \"error\": true,\n        \"message\": \"错误消息\",\n\
    \        \"detail\": {}\n    }\n    ```\n    "
  version: 0.1.0
paths:
...
  /api/upload/production/local:
    post:
      tags:
      - upload
      summary: 本地方式上传生产信息
      description: 通过表单上传 Excel 文件到本地存储，并解析保存到数据库。需要 Bearer Token 认证。
      operationId: upload_production_local_api_upload_production_local_post
      security:
      - HTTPBearer: []
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_production_local_api_upload_production_local_post'
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema: {}
              example:
                message: 上传成功
                filename: production_info.xlsx
                saved_path: uploads/production_info_20250101_123456.xlsx
                records_processed: 150
        '400':
          description: 文件类型不支持或文件格式错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/upload/production/oss:
    post:
      tags:
      - upload
      summary: OSS 方式上传生产信息
      description: 处理已上传到阿里云 OSS 的生产信息 Excel 文件，下载并解析保存到数据库。需要 Bearer Token 认证。
      operationId: upload_production_oss_api_upload_production_oss_post
      security:
      - HTTPBearer: []
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductionOSSUploadRequest'
      responses:
        '200':
          description: 处理成功
          content:
            application/json:
              schema: {}
              example:
                message: 处理成功
                object_key: uploads/production/2025/01/production_info.xlsx
                records_processed: 150
        '400':
          description: OSS 文件不存在或文件格式错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/upload/worklog/local:
    post:
      tags:
      - upload
      summary: 本地方式上传员工工作量
      description: 通过表单上传 Excel 文件到本地存储，并解析保存到数据库。需要 Bearer Token 认证。
      operationId: upload_worklog_local_api_upload_worklog_local_post
      security:
      - HTTPBearer: []
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_worklog_local_api_upload_worklog_local_post'
      responses:
        '200':
          description: 上传成功
          content:
            application/json:
              schema: {}
              example:
                message: 上传成功
                filename: worklog_202501.xlsx
                saved_path: uploads/worklog_202501_20250101_123456.xlsx
                records_processed: 200
        '400':
          description: 文件类型不支持或文件格式错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/upload/worklog/oss:
    post:
      tags:
      - upload
      summary: OSS 方式上传员工工作量
      description: 处理已上传到阿里云 OSS 的员工工作量 Excel 文件，下载并解析保存到数据库。需要 Bearer Token 认证。
      operationId: upload_worklog_oss_api_upload_worklog_oss_post
      security:
      - HTTPBearer: []
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorklogOSSUploadRequest'
      responses:
        '200':
          description: 处理成功
          content:
            application/json:
              schema: {}
              example:
                message: 处理成功
                object_key: uploads/worklog/2025/01/worklog_202501.xlsx
                records_processed: 200
        '400':
          description: OSS 文件不存在或文件格式错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/upload/oss/presigned-url:
    get:
      tags:
      - upload
      summary: 获取 OSS 预签名 URL
      description: 生成阿里云 OSS 的预签名 URL，用于前端直接上传文件到 OSS。需要 Bearer Token 认证。
      operationId: get_presigned_url_api_upload_oss_presigned_url_get
      security:
      - HTTPBearer: []
      parameters:
      - name: object_key
        in: query
        required: true
        schema:
          type: string
          description: OSS 对象键(文件路径)
          title: Object Key
        description: OSS 对象键(文件路径)
        example: uploads/production/2025/01/file.xlsx
      - name: expires
        in: query
        required: false
        schema:
          type: integer
          maximum: 86400
          minimum: 1
          description: 过期时间(秒),范围 1-86400，默认 1 小时
          default: 3600
          title: Expires
        description: 过期时间(秒),范围 1-86400，默认 1 小时
        example: 3600
      - name: method
        in: query
        required: false
        schema:
          type: string
          description: HTTP 方法,默认 PUT
          default: PUT
          title: Method
        description: HTTP 方法,默认 PUT
        example: PUT
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema: {}
              example:
                presigned_url: https://bucket.oss-cn-hangzhou.aliyuncs.com/uploads/file.xlsx?Expires=1234567890&OSSAccessKeyId=xxx&Signature=xxx
                object_key: uploads/file.xlsx
                expires: 3600
                method: PUT
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationErrorResponse'
        '422':
          description: 请求参数验证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorResponse'
...
components:
  schemas:
    AuthenticationErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: AuthenticationErrorResponse
      description: 认证错误响应模型
      example:
        detail: {}
        error: true
        message: 认证失败
    Body_upload_production_local_api_upload_production_local_post:
      properties:
        file:
          type: string
          format: binary
          title: File
          description: Excel 文件（.xlsx 或 .xls 格式）
        filter_month:
          anyOf:
          - type: string
          - type: 'null'
          title: Filter Month
          description: 月份过滤,格式:yyyyMM
      type: object
      required:
      - file
      title: Body_upload_production_local_api_upload_production_local_post
    Body_upload_worklog_local_api_upload_worklog_local_post:
      properties:
        file:
          type: string
          format: binary
          title: File
          description: Excel 文件（.xlsx 或 .xls 格式）
      type: object
      required:
      - file
      title: Body_upload_worklog_local_api_upload_worklog_local_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InternalServerErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: InternalServerErrorResponse
      description: 服务器内部错误响应模型
      example:
        detail:
          message: Connection timeout
          type: DatabaseError
        error: true
        message: 服务器内部错误
...
    ProductionOSSUploadRequest:
      properties:
        object_key:
          type: string
          title: Object Key
          description: OSS 对象键(文件路径)
          example: uploads/production/2025/01/production_info.xlsx
        filter_month:
          anyOf:
          - type: string
          - type: 'null'
          title: Filter Month
          description: 月份过滤,格式:yyyyMM
          example: '202501'
      type: object
      required:
      - object_key
      title: ProductionOSSUploadRequest
      description: 生产信息 OSS 上传请求
      example:
        filter_month: '202501'
        object_key: uploads/production/2025/01/production_info.xlsx
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ValidationErrorResponse:
      properties:
        error:
          type: boolean
          title: Error
          description: 是否为错误响应
          default: true
        message:
          type: string
          title: Message
          description: 错误消息
        detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Detail
          description: 错误详情
      type: object
      required:
      - message
      title: ValidationErrorResponse
      description: 验证错误响应模型
      example:
        detail:
          field: order_no
          reason: 订单号不能为空
        error: true
        message: 数据验证失败
...
    WorklogOSSUploadRequest:
      properties:
        object_key:
          type: string
          title: Object Key
          description: OSS 对象键(文件路径)
          example: uploads/worklog/2025/01/worklog_202501.xlsx
      type: object
      required:
      - object_key
      title: WorklogOSSUploadRequest
      description: 员工工作量 OSS 上传请求
      example:
        object_key: uploads/worklog/2025/01/worklog_202501.xlsx
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 输入 JWT token，格式：Bearer <token>
...